Add files via upload
Made a lot of updates, wanted to make a new folder for simplicity. The general framework is there. Good number of errors unfortunately, but I do believe that once they are fixed, this should work for our project. The two things that are not implemented yet are the graphics and the learning, both of which I would imagine to be more straightforward than the actual problem. I think it would be great if we could get this done before Thursday.